requirements.txt
streamlit
pandas
Pillow
piexif
openai
iptcinfo3